Removed catan game. Moved tests to catan game engine
