--- src/io/github/notaphplover/catan/core/game/ICatanGame.java
@@ -1,23 +1,0 @@
-package io.github.notaphplover.catan.core.game;
-
-import io.github.notaphplover.catan.core.board.ICatanBoard;
-import io.github.notaphplover.catan.core.player.IPlayer;
-import java.util.Map;
-
-public interface ICatanGame {
-  IPlayer getActivePlayer();
-
-  ICatanBoard getBoard();
-
-  IPlayer[] getPlayers();
-
-  Map<IPlayer, Integer> getPoints();
-
-  int getPointsToWin();
-
-  GameState getState();
-
-  int getTurnNumber();
-
-  boolean isTurnStarted();
-}
